Update cicd github workflow and run local and remote tests together
